+++ src/main/java/org/nefedov/weather/application/dto/LocationDto.java
@@ -1,4 +1,6 @@
 package org.nefedov.weather.application.dto;
 
-public record LocationDto(String name, Double latitude, Double longitude) {
+import java.math.BigDecimal;
+
+public record LocationDto(String name, BigDecimal latitude, BigDecimal longitude) {
 }
